added necessary modifications
